Move user interaction and upstream.yml routines to separate modules.
