+++ Lib/gftools/packager/exceptions.py
@@ -1,0 +1,8 @@
+class UserAbortError(Exception):
+    pass
+
+
+class ProgramAbortError(Exception):
+    pass
+
+
